Rename: organisms 폴더 안에 내용물 pages 폴더로 옮기기
