--- src/components/organisms/mypage/index.js
@@ -1,3 +1,0 @@
-export { default as MyArticleBoard } from './MyArticleBoard';
-export { default as MyArticlePreview } from './MyArticlePreview';
-export { default as MyProfile } from './MyProfile';
