+++ src/components/pages/Mypage/MyArticleBoard/index.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-import { UserService } from '../../../network';
+import { UserService } from 'network';
 import PreviewArticle from '../category/PreviewArticle';
 import MyArticlePageSelector from './MyArticlePageSelector';
