+++ src/components/pages/Mypage/MyProfile/LinkSection/IconSet.jsx
@@ -1,5 +1,3 @@
-import styled from 'styled-components';
-
 const IconGithub = () => {
   return (
     <svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
@@ -73,26 +71,6 @@ const IconSolvedAC = () => {
   );
 };
 
-const MypageButton = styled.button`
-  padding: 0.1rem 1rem;
-  margin: 0.1rem;
-  border: none;
-  width: 6rem;
-  height: 1.6rem;
-  font-size: 0.8rem;
-  white-space: nowrap;
-  background-color: ${props =>
-    props.btnType === 'auth-42'
-      ? props.theme.primary
-      : props.btnType === 'auth-42-done'
-      ? props.theme.buttonGray1
-      : props.theme.secondary};
-  &:hover {
-    cursor: ${props =>
-      props.btnType === 'auth-42-done' ? 'default' : 'pointer'};
-  }
-  color: ${props => props.theme.white};
-`;
 
 const IconSet = {
   Icon42,
@@ -102,4 +80,5 @@ const IconSet = {
   IconTwitter,
   IconSolvedAC,
 };
-export { IconSet, MypageButton };
+
+export default IconSet
